LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/last-successful-change-fallbacks-position-area-expected.txt


PASS Starts rendering with fallback
PASS Both base and fallback works, position-try-fallbacks is set to the same value, keeps fallback
PASS position-try-fallbacks is set to a different value, both base and fallback works, uses base because last successful position option is invalidated
